fix(api): updates to handle session timeout

+++ example/testpod-stop.sh
@@ -16,25 +16,33 @@ runCommand "kubectl delete role external-provisioner-cfg-systems"
 runCommand "kubectl delete rolebinding csi-provisioner-role-cfg-systems"
 
 runCommand "helm uninstall test-release"
-runCommand "kubectl delete pods test-pod --grace-period=0 --force"
 runCommand "kubectl delete secrets seagate-exos-x-csi-secrets"
 runCommand "kubectl delete sc systems-storageclass"
-runCommand "kubectl delete pvc systems-pvc"
+
+pv=$(kubectl get pv | grep pv | awk '{print $1}')
+if [ ! -z "$pv" ]; then
+    kubectl patch pv $pv -p '{"metadata": {"finalizers": null}}'
+    runCommand "kubectl delete pv $pv --grace-period=0 --force"
+fi
 
 pvc=$(kubectl get pv | grep pvc | awk '{print $1}')
-runCommand "kubectl patch pv $pvc -p '{"metadata": {"finalizers": null}}'"
-runCommand "kubectl delete pv $pvc --grace-period=0 --force"
+if [ ! -z "$pvc" ]; then
+    kubectl patch pv $pvc -p '{"metadata": {"finalizers": null}}'
+    runCommand "kubectl delete pv $pvc --grace-period=0 --force"
+fi
 
-runCommand "kubectl delete pods --all"
+pod=$(kubectl get pod | grep seagate-exos-x-csi-controller | awk '{print $1}')
+if [ ! -z "$pod" ]; then
+    runCommand "kubectl delete pod $pod --grace-period=0 --force"
+fi
+
+pod=$(kubectl get pod | grep seagate-exos-x-csi-node | awk '{print $1}')
+if [ ! -z "$pod" ]; then
+    runCommand "kubectl delete pod $pod --grace-period=0 --force"
+fi
+
+runCommand "kubectl delete pod test-pod --grace-period=0 --force"
 
 banner "Check Resources"
 
-runCommand "kubectl get configmaps"
-runCommand "kubectl get serviceaccounts"
-runCommand "kubectl get daemonsets"
-runCommand "kubectl get deployments"
-runCommand "kubectl get pvc"
-runCommand "kubectl get pvc"
-runCommand "kubectl get sc"
-runCommand "kubectl get secrets"
-runCommand "kubectl get pods"
+runCommand "kubectl get all"
